+++ src/main/org/firebirdsql/management/FBServiceManager.java
@@ -22,10 +22,7 @@ import org.firebirdsql.gds.ServiceRequestBuffer;
 import org.firebirdsql.gds.impl.GDSFactory;
 import org.firebirdsql.gds.impl.GDSServerVersion;
 import org.firebirdsql.gds.impl.GDSType;
-import org.firebirdsql.gds.ng.FbDatabaseFactory;
-import org.firebirdsql.gds.ng.FbService;
-import org.firebirdsql.gds.ng.FbServiceProperties;
-import org.firebirdsql.gds.ng.IServiceProperties;
+import org.firebirdsql.gds.ng.*;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -197,6 +194,21 @@ public class FBServiceManager implements ServiceManager {
         FbService fbService = dbFactory.serviceConnect(serviceProperties);
         fbService.attach();
         return fbService;
+    }
+
+    protected FbDatabase attachDatabase() throws SQLException {
+        if (database == null) {
+            throw new SQLException("Property database needs to be set.");
+        }
+        FbConnectionProperties connectionProperties = new FbConnectionProperties();
+        connectionProperties.setServerName(serviceProperties.getServerName());
+        connectionProperties.setPortNumber(serviceProperties.getPortNumber());
+        connectionProperties.setUser(serviceProperties.getUser());
+        connectionProperties.setPassword(serviceProperties.getPassword());
+        connectionProperties.setDatabaseName(database);
+        FbDatabase fbDatabase = dbFactory.connect(connectionProperties);
+        fbDatabase.attach();
+        return fbDatabase;
     }
 
     public void queueService(FbService service) throws SQLException, IOException {